Added support for running from Node/browser
