+++ package.json
@@ -6,7 +6,7 @@
   "homepage": "https://bitwave.tv",
   "author": "dvdrw",
   "license": "LGPL-3.0",
-  "main": "api.js",
+  "main": "api.mjs",
   "scripts": {
     "test": "echo \"TODO: testing. Nothing to see here. Move along...\""
   },
